FEM: import tools, move fill result obj with imported results inside import tools module
